+++ xtp-tutorials/pyxtp/scripts/run_optimization.py
@@ -12,6 +12,8 @@ calc = xtp(nthreads=2)
 calc.select_force(energy='singlets', level=0, dynamic=False)
 atoms.calc = calc
 
+atoms.get_forces()
+
 dyn = QuasiNewton(atoms, trajectory='test.traj')
 dyn.run(fmax=0.01)
 write('final.xyz', atoms)